+++ src/types/options/index.ts
@@ -10,3 +10,4 @@ export * from './LoadingIndicatorOptions';
 export * from './LoaderOptions';
 export * from './PresetOptions';
 export * from './UIOptions';
+export * from './SelectionOptions';